requirements.txt
gunicorn==2.4.1
line-bot-sdk==1.16.0
bleach==3.1.4
